main/mandoc: detect at run-time if the pager supports -T
This is useful to ensure that we support tag files with mandoc
if main/less is installed (which it will be on most systems).
Tag file support in mandoc is _extremely_ useful as it allows
one to, for example, directly lookup options in a man page.

This patch is (hopefully) upstreamable.

Alternative solutions:

* Implement the -T flag in BusyBox less
* Add a runtime dependency on main/less to mandoc
* Make BusyBox less ignore the -T flag