testing/lemmy-ui: hardcode translation commit
Even though having them automatically in sync is nice, if you would
change the commit in lemmy, you need to bump the pkgrel in lemmy-ui
anyway to make sure it's rebuilt.

Besides, tools like lua-aports start to fail because `$startdir` is not
defined when it sources the packages, resulting in errors.