testing/gpsbabel: Assuming maintainership and enable older architectures
Due to the "cutting edge" dependency on qt6 and qt6-qt5compat only 64-bit
architectures were enabled as otherwise the dependency would be
missing.

Reached out and confirmed with the maintainer of the qt6 and qt5
packages (Bart Ribbers <bribbers@disroot.org>), that building those on
32-bit, although not impossible, comes with its own problems, and is out
of scope for now.

Tested on armv7, not tested on x86, as I have no access, but that'd
probably work too.