testing/pounce: add supervise-daemon to conf.d
The pounce IRC bouncer runs as a foreground process by default and
crashes if it loses connection to an IRC network. The calico program
also runs in the foreground. Both of these programs don't have options
to fork to the background. Using supervise-daemon ensures that they can
be respawned, if needed.