main/linux-lts: fix x86 PAE and HIGHMEM
Fix issue with only having 849MB memory for 32 bit x86 kernels.

Revert enabling PAE of lts. This solves problem with older hardward that
no longer can boot due to PAE enabled in kernel. This partially reverts
commit ea95865daefe (main/linux-lts: enable PAE and XEN for 32 bit x86)

ref: https://lists.alpinelinux.org/~alpine/devel/%3C84d45683-3727-8e6e-2f65-04c5f724f0da%40dkcomm.net%3E

We keep PAE and XEN for the virt kernel, and bump it to 686. The reason
for enabling PAE was to be able to run XEN PVs, so this can now be done
with the virt kernel. We also assume that any VMs run on x86_64
hardware, since it does not make any sense to run a 32 bit VM host.

Also use HIGHMEM64G for virt kernel so we can use more than 4G there.

ref: https://gitlab.alpinelinux.org/alpine/aports/-/commit/ea95865daefe76f6fa92806f747e98e829320d32#note_330191