community/linux-edge: add initial support for pinebook pro
This commit enables a minimal set of config options to support (most of)
the hardware in the pinebook pro. There are known missing config options
to add support for certain features at this time but I haven't been able
to figure out which ones are needed yet (they are working on my local
configuration[0] though). The following are known issues:

* missing temperature sensor support
* PCIe errors in dmesg

[0] https://git.sr.ht/~svmhdvn/dotfiles/tree/master/item/kernel_configs/pbp_linux_6.5.3.config

Tested using the default u-boot package in alpine-edge with EFI boot.