community/clamav: fix clamd.conf
New release of clamav has changed $srcdir/etc/clamd.conf.sample such that there
are now two commented out LocalSocket example lines. This causes the sed in
APKBUILD's package function to uncomment both. This causes the clamd.initd
start_pre functiion to emit a two line value for the local var socket, in turn,
causing the evaluation of socketdir=${socket%/*} to be the whole first line.
Now, when clamd.initd's start_pre check_path of $socketdir runs, it creates
/run/clamav/clamd.sock as a directory, and now the start function cannot
remove it.

This replaces the sed statements in package() function with a patch.
Using a patch for this seems appropriate to me because the patch can be adjusted
at later upgrades, but the sed statements can do unexpected things when the
target file changes.