community/kde-cli-tools: rework workaround for a circular dependency
To fix a circular dependency https://gitlab.alpinelinux.org/alpine/aports/-/issues/11785
we previously let kde-cli-tools install itself when plasma-workspace was
installed through a install_if. However that means that if there are 2
versions of kde-cli-tools available in the enabled repositories, in my
case a Qt5 and a Qt6 one, the "old" one will keep pulling itself in
which blocks installation of plasma-workspace (due to requiring both KF5
and KF6 packages at the same time, conflicting with each other).

plasma-workspace still depends on kde-cli-tools at runtime and kde-cli-tools
still requires plasma-workspace for building, but now we just pull the
dep in through plasma-desktop, plasma-mobile and plasma-bigscreen to fix
this

Note that this circular dependency is resolved in the Qt6 builds so this
won't be a problem in the future