community/clojure: enable on 32-bit (armhf, armv7, x86) with OpenJDK 8 LTS
Clojure actually requires openjdk version 8 as a
minimum (https://clojure.org/releases/downloads#_java_compatibility),
and that is still a supported LTS release.

Incidentally that's the last version with official 32-bit support,
and the only version available in the Alpine repos for 32-bit
architectures.

All in all, it makes sense to align the arch field for openjdk8 and
clojure.