community/gnome-session: force bash login shell on workaround
See https://bugzilla.gnome.org/show_bug.cgi?id=736660 for context (you
can probably scroll to the end of the discussion)

In alpine, we compile bash as-is, without
"NON_INTERACTIVE_LOGIN_SHELLS", which is what makes it work in other
distros. Forcing the non-POSIX compliant variation when we know we're
bash should actually work.