community/nerd-fonts: restore nerd fonts
Upstream nerd fonts ships:

- The nerd font symbols themselves.
- Several other fonts patched to include nerd font symbols.

The upgrade to upgrade to 2.3.3 was a full rewrite of this package, and
dropped font-nerd-fonts-symbols; only patched versions of other fonts
were retained. I believe this drop was done inadvertently during the
refactor and not intentional.

Previously the nerd font symbols font came in two variants,
font-nerd-fonts-symbols-2048em and font-nerd-fonts-symbols-1000em.
As of v3.0.0, both variants have been merged into one. This commit
re-introduces nerd font symbols as "font-nerd-fonts-symbols".

A replaces= directive is also included so that existing installations of
the previous version will upgrade to this one (upstream has dropped the
two variations that previously existed and now only ships a single one).

This also includes the upstream-provided fontconfig file.

Fixes: https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/38362