Revert "community/startup-notification: fix 32-bit time_t"
Most projects linking to startup-notification uses a pointer `long`
instead of a pointer to `time_t`. The compiler is not able to pick this
up so on 32 bit arches it results in a crash.

Revert to use `long` to solve this, til upstream has a better longterm
solution.

ref: https://gitlab.alpinelinux.org/alpine/aports/-/issues/15441

This reverts commit e99c0e1ab1248b67d25763b493b9a9b8413dc74b.

(cherry picked from commit 81834c57c29a06cd4f3b4a5750dadaf9f63f00ff)