community/tepl: drop maintainership
tepl starting from 6.5.1 started depending on libgedit-gtksourceview
and from 6.8.0 on libgedit-amtk-5.
for a tepl 6.8.0 and a gedit/gnome-latex upgrade both of these forks
need to be packaged. i don't want to, but someone else can pick them up
if they want.

ripgrepping on aports resulted in only gedit and gnome-latex depending
on tepl