main/sngtc_client: fix license expression
As far as I can tell, this aport builds three different targets, each of
which is under a different license. One is under a butchered
BSD-3-Clause that doesn't have a SPDX identifier (thus specified as
custom), one is a standard GPL-2.0-or-later, and one is under the
standard BSD-3-Clause.