main/musl: backport patch to remove basename prototype from string.h
In c23 mode, the basename declaration from string.h is incompatible
with the one provided in libgen.h. This causes issues with software
that includes both libgen.h and string.h with -D_GNU_SOURCE in c23
mode. For example, community/tofi. Short of removing the prototype
there is no other good workaround for this.