community/openjdk8: remove close-fd workaround again
Remove this workaroudn and use opendir() on /proc/self/fd again.
This fixes an issue on our internal build server, where the openfiles ulimit was at 1bn, causing a single Runtime.exec() to take over 9 seconds. This patch should not be needed any longer, as the java 9+ aoirts didn't include it in the first place and seem to run just fine.