main/mkinitfs: upgrade to 3.9.0_rc2, refactor trigger
Let the trigger look for /lib/modules/* so we run the trigger when third
party modules (eg zfs) are updated. The third party modules may be used
in initramfs.

This solves the issue when 3rd party module (zfs) used by initramfs was
updated, but initramfs was not regenerated and caused system to be
unbootable.