main/mkinitfs: fix trigger for upgrades
When we upgrade we also get the directory for the old version as an arg.
We must handle this properly and keep the initramfs (which may be
updated in previous iterations), while also skipping the directories for
the old version.

Detect this by look for the modules.order file. If its missing, assume
we are upgrading.

fixes: https://gitlab.alpinelinux.org/alpine/aports/-/issues/15538