community/helix: use system queries
For a while I have been using this personally, and I only remembered it
recently when I couldn't access some system queries on a remote. It is
important for queries and grammars to be in sync, however while this
fixes a lot of problems, it causes a few new ones aswell, but it is
generally a net gain.

The helix folk are (somewhat) coordinating with neovim so that their
interpretations of grammars match, so that queries should be easier to
upstream.