community/avr-libc: update I/O headers from vendor toolchain
Both the avr-libc in the latest release and in the current HEAD in the
repo the vendor header files are pretty outdated. They are missing
I/O descriptions for more recently released MCUs as well as
headers with missing types and defines for e.g. ATXmega1281U. Updating
them fixes the issue.