testing/perl-uri-redis: move perl-test-most to makedepends
This fixes the build on riscv64, as it tries to
autoinstall the Test::Most it deems missing due
to tests being disabled on the riscv64 builder,
which means checkdepends aren't installed.

The alternative would involve patching Makefile.PL
to remove the autoinstall directive, but it doesn't
hurt to have perl-test-most in makedepends, so this
is the solution being chosen for the moment.