testing/py3-litex-hub-modules: avoid global exec
Shelling out in the global scope is expensive and in this case not
necessary, change the `cut` invocation to parameter substitution.

`${foo%:*}` cuts of the `:` and everything that comes after it, which is
the same as `| cut -d : -f 1` does.

`${foo#*:}` cuts of everything until and including the `:`, which is
the same as `| cut -d : -f 2` does.