community/waydroid: enable nftables in the net script
this is also enabled in distros like fedora. it was disabled by
default some time ago. (starting from 1.2.0)

this fixes waydroid not able to run the networking script on pmOS
devices, as those kernels don't ship the IPTables/NFTables compat
layer. i don't think we should be relying on that anyway, so yeah.