main/grub: remove fwsetup check for efi systems
This prevents machines from booting if they are still running versions
of the grub bootloader while the config is updated to 2.12.

Showing a menu entry for uefi-firmware that may not work is infinitely
better than crashing and preventing the machine to boot, so remove the
condition and always show the menu entry.

See also: https://bugs.archlinux.org/task/75701