community/phosh-osk-stub: move from testing
Guido gave the green light on switching the default on-screen keyboard
for postmarketOS Phosh from Squeekboard to phosh-osk-stub due to the
former effectively being unmaintained for soon a year by now. As such,
we need to move phosh-osk-stub to community so that it will be available
in stable releases as well.