Revert "community/lxd: add depend nftables"
Do not enforce the use of nftables.

When installed nftables is used. While this has advantages over
iptables for example with the use of namespaces which is less
dependant on rules orders, you might still end up with a
non-working LXD network if there's a table that blocks protocols
and ports required by LXD. I belive that the proper fix should be
LXD setting a prority on its table. For now we just remove the
nftables dependency and let the user install it if needed. This
will cause LXD to log this warning upon startup:

  Firewall failed to detect any compatible driver, falling back to \"xtables\"

https://documentation.ubuntu.com/lxd/en/stable-5.0/howto/network_bridge_firewalld/#xtables-vs-nftables

This reverts commit 5d660cfe1708fb429a5ed0d755627e8ed4eab70d.