testing/incus: do not enforce the use of nftables
When installed nftables is used. While this has advantages over
iptables for example with the use of namespaces which is less
dependant on rules orders, you might still end up with a
non-working Incus network if there's a table that blocks protocols
and ports required by Incus. I belive that the proper fix should be
Incus setting a prority on its table. For now we just remove the
nftables dependency and let the user install it if needed. This
will cause Incus to log this warning upon startup:

  Firewall failed to detect any compatible driver, falling back to "xtables"

https://linuxcontainers.org/incus/docs/main/howto/network_bridge_firewalld/#xtables-vs-nftables