testing/py3-litex-hub-modules: fix circular dep
The main package is empty and was intended as a virtual meta-package
that pulls in all the subpackages. However, because the dependencies
were listed in the global scope, they applied to all subpackages as
well. This moves the dependencies to the package function.

As result it is now be possible to install just a single
subpackage without pulling in all the other modules.

Fixes https://gitlab.alpinelinux.org/alpine/aports/-/issues/15820