community/kde-frameworks: downgrade *5 packages back to latest non-KDE6 release
The 6.0.0 release of these packages don't actually build with Qt6, but
we still need to keep them around for some stuff like Breeze. So for now
just keep them at their last non-KDE6 release.

This also sets the group tag to kde-frameworks-old so we can work with
them independently in scripts