main/openssh: remove fix-verify-dns-segfault.patch
This patch was added to fix #8323 for Alpine 3.10. As it turns out,
this was actually a bug/incompatibility in musl, as musl did not
set h_errno and therefore did not indicate that the requested data
was not found in the DNS response.

Nowadays, musl sets h_errno and therefore the getrrsetbyname() compact
function correctly recognizes that the data is not available in the DNS
response and does not perform a null pointer dereference even without
this patch.

Specifically, this is fixed by the following commit in the musl
repository: https://git.musl-libc.org/cgit/musl/commit/?id=19f8642494b7d27b2ceed5c14d4a0b27cb749afe
This commit is part of musl 1.2.2 and since this version the segfault
does not occur without the fix-verify-dns-segfault.patch on Alpine. I
went the extra mile of verifying this by installing Alpine 3.9 (which
didn't inculde fix-verify-dns-segfault.patch yet) in QEMU, checking that
the segfault occurs there and conforming that updating musl to backport
this patch fixes the segfault.