community/zeal: change source to release's git tag archive
The release archives posted on github include a zeal.desktop file with
the wrong line endings, reported upstream here:
https://github.com/zealdocs/zeal/issues/1610

Switching the source archive to the release's git tag avoids the this
problem since it builds from what's in the git repo at that tag. The
file has the correct line endings in the git repo, it's only the release
source archive where this is wrong.

This also forces the project to be built with QT5. Even though upstream
claims it could be built with QT6, cmake fails to detect things
correctly (ECM pulls in both qt5 and qt6, and confuses it, it seems)