community/texmf-dist: upgrade to 2024.0
This update is minor from the user perspective, but a major change to
the APKBUILD.

The background is that before we relied on Arch Linux already bundling
texlive packages into collections, that we based the subpackages on.
This was an easy way to reduce the maintenance offered on our side.
However, Arch Linux stopped doing that, so this approach no longer
works.

The new approach is to download the texmf dist bundle from the tug
archives and parse the texlive.tlpdb database with a few awk commands to
extract which collections are provided, which texlive pkgs they consists
of, and which file those in term contain. I believe that this will have
an even lower maintenance effort while also giving us independence from
Arch Linux.