community/go: copy 'misc' directory to $GOROOT
Some applications hard code the 'misc' directory and expect it to be
present in $GOROOT Other distributions (e.g. Arch) also install it in
this path. Presently, the the 'misc' directory is mistakenly added to
go-doc.

While at it, the copy command for copying the 'src' directory can be
simply added to the existing command which copies the directories 'pkg'
and 'lib.

Fixes !63444

Previously:

* #15008
* !47919
* 03fe60c18472cb133b775c3a8600e3b0b661c704