community/texmf-dist: provide subpackages at global scope
Don't extend $subpackages at runtime, but define it at global scope.

Sadly, this means manual effort to keep the subpackages in sync
with the collections they pack. A check() function was added to check
if for every collection a matching subpackage is defined. (If there is
a subpackage without a corresponding collection, this will fail
when packing anyway and is not explicitly tested.)