testing/libhandy: move from community
The only remaining user of libhandy(0) is in testing. All other users
have migrated to libhandy1. As such, move libhandy(0) itself to testing
too so we don't need to include this library in stable releases at
least given that my MR to remove libhandy(0) entirely[1] is in a state
of limbo.

 [1]: https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/63787