main/samba: ensure subpackage versions are in sync
Samba's .so versioning is not enough to prevent version mismatches. That
can cause missing symbol errors.

To keep dependencies simple most packages depend on samba-libs with a
strict version. And samba-libs depends on it's dependencies with a
strict version. Other packages are to my knowledge not effected by
version mismatches.

Maybe fixes [aports#15818] and [apk-tools#10999]?

[aports#15818]: https://gitlab.alpinelinux.org/alpine/aports/-/issues/15818
[apk-tools#10999]: https://gitlab.alpinelinux.org/alpine/apk-tools/-/issues/10999