community/texmf-dist: drop duplicate man pages
Some man pages were provided already by texlive-doc. It indeed appears
that the git repo for the texlive binaries and the texmf-dist tree
contain duplicates; this became apparent only because of the
recent update that moves the man pages of texmf-dist where man expects
them.

For now: Use a manually managed list of duplicates and remove them.
At least there is only one release of texmf-dist per year ...

The reason to prefer the man pages in texlive-doc over the one in
texmf-dist-doc is that building the doc from the same repo as the
binaries appears to be more likely to have them in sync.