community/texmf-dist: fix -doc subpackages
Before I naively used `default_doc` for `-doc` subpackages, but this
only works if there is only a single `-doc` subpackages, as otherwise
all `-doc` packages would be installed if `docs` and the root package
is installed.

This fixes the issue by hand-rolling the split function for the
non-root `-doc` subpackage.