community/texmf-dist: allow installing duplicate man pages
This tries to be a maintainable solution for the issue that both the
source repo for the texlive package and the sources for texmf-dist
contain man pages for the applications provided by the texlive
package.

It adds a postfix to the man pages of texmf-dist, so that they can
be installed at the same time. This wastes storage space for the
duplicates. Especially since the man pages provided by `texlive-doc`
should be considered the single source of truth for the man pages, it
is suboptimal to have the duplicates. However, is absence of any
obvious clean way to remove the duplicates, this may be a sensible
alternative.