community/gnome-session: enable systemd session support
This will be the new default with GNOME Session 46. While it may sound
like something that would break non-systemd support, in reality what
happens is that GNOME Session fails to find systemd and falls back to
the new behaviour. This is being enabled separately from the GNOME
Session 46 upgrade to get this a chance of being tested more widely
before upgrading.