community/py3-gitlab: move from testing
We use this for CI downstream in postmarketOS and it would be useful to
have this in community so we in the future can use stable releases for
that purpose. Additionally, this package has been in testing for almost
two years and has been actively maintained since. As such, I think it
makes sense to move this to community.

It was also previously in community, but removed due to a lack of users:
https://gitlab.alpinelinux.org/alpine/aports/-/commit/f2c7b02b10febfdce0e6ac2f0e696f425db88043

However, this is clearly not the case any more.