community/texmf-dist: let texmf-dist-bibtexextra provide biblatex
This fixes a conflict between biblatex and
texmf-dist-bibtexextra-{doc,examples}. It prevents installing both
biblatex and `texmf-dist-bibtexextra` at the same time, which otherwise
could result in the package `biblatex` overwriting the `biblatex`
version in `texmf-dist-bibtexextra` with something that is incompatible
with the other TeXLive packages (e.g. the various biblatex styles).