community/megapixels: upgrade to 1.8.1
This also adds a patch to not segfault when taking pictures. The underlying
causes is a breaking change in libtiff. This patch is needed with at least
libtiff 4.6.0t that is currently in aports master, though possibly it has been
needed since libtiff 4.5.1. In either case, Megapixels upstream is considering
a more permanent fix for these libtiff backward-incompatibilities,
so the next release which has that fix can drop this patch.

Ref: https://gitlab.com/megapixels-org/Megapixels/-/issues/4
Ref: https://gitlab.com/postmarketOS/pmaports/-/issues/2764