community/sway: bring -portalconf up to date with xdg
`DesktopNames` is now part of the spec, so remove the `X-` prefix,
as gdm only queries `DesktopNames`. This actually sets XDG_CURRENT_DESKTOP
with gdm now.

Also add a warning for users of sway-portalsconf, that not all portals
are provided by us. For things like the keyring or color-scheme more
portal backends have to be specified.

Also move the `sway-portals.conf` to `/usr/share/...` instead of `/etc`.