community/gnome-control-center: enable location panel
Was disabled by default upstream as a reaction to the announcement of
the Mozilla Location Service shutdown. The goal was to give distros an
option to decide based on their own release schedule and support
lifetime. Since then, it's been reenabled, and we have a fast
development cycle, so let's enable it.