community/dillo: upgrade to 3.1.0_git20240506
This switches `source` from https://github.com/w00fpack/dilloNG
to https://github.com/dillo-browser/dillo which is a
successor to the original project.

Building the git revision so there is a change in `pkgver`
to indicate that `source` has changed.

The HTTPS backend has also been switch to OpenSSL, as
MbedTLS 3.6.0 now defaults to TLSv1.3 as max version,
and that is not currently supported by the code here.