main/dash: fix install of -binsh
When installing dash-binsh but dash is not installed the previous binsh
may be removed early leaving dash.post-install without a shell to run.
Using the shell itself also resolves a recursive dependency between dash
and dash-binsh.

Additional dependency on busybox is needed when creating a rootfs to
ensure the 'add-shell' command is actually available when running the
dash.post-install script.