main/yash: fix install of -binsh
When installing yash-binsh but yash is not installed the previous binsh
may be removed early leaving yash.post-install without a shell to run.
Using the shell itself also resolves a recursive dependency between yash
and yash-binsh.

Additional dependency on busybox is needed when creating a rootfs to
ensure the 'add-shell' command is actually available when running the
yash.post-install script.