community/neo4j: ensure JDK11 is used by setting JAVA_HOME
This is the last blocker for build-3-20-aarch64. I tried
building it in CI (without this change), and it succeeded,
so i suspect that there may be another version of JDK
installed on the builder, which is used instead of JDK11
(JDK11 is required by Neo4j for successful building).