testing/gdb-esp: drop package
The package caused GDB to request reading of non-existing register from
OpenOCD when single-stepping through the code, resulting in OpenCOD
closing the GDB connection. It seems that only the xtensa-dynconfig
plugin approach to configuring the xtensa CPU parameters works with
recent versions of GDB.

Since this has been broken for some time, it is unlikely to be missed.
https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/66926
also provides the xtensa-dynconfig binutils/GCC/GDB plugins and a
wrapper for GDB for ease of use, so that an upgrade path is
on the horizon.