community/biber2.19: work around limitation of lua-aports
When sourcing the APKBUILD of texmf-dist, which provides biblatex~3.19
in the subpackage function, this info is not available when determining
the order in which to build packages for a release. This works around
this by adding texmf-dist-bibtexextra as build dependency and only
adding biblatex~3.19 in the package function to the dependencies.