community/texlive: add missing dependency
The base package already provides LaTeX and not only TeX, so we also
need to install texmf-dist-latex.

In theory, it would be possible to provide only TeX support and only
need texmf-dist without texmf-dist-latex. But in practice most people
will use LaTeX over plain old TeX and e.g. pdflatex is only a symlink
to pdftex, so there is not much storage to be saved anyway.