community/texlive: fix trigger
The hyphenation files where in older versions of texmf-dist always
available, but upstream split them into the lang* collections.

Now that each texfmf-dist-lang* provides the chunk of the language
configuration it provides the hyphenation data for, we can simply
do `cat lang.d/*.dat > language.dat` to generate the monolithic one
matching the installed set of languages.

In addition, the output of the trigger was improved to make it more
obvious something fails. A hint is given that the we do not disable
the generation of uninstalled formats, as fmtutil handles missing
formats gracefully and this is easier to maintain.