main/wasi-libcxx: use --target=wasm32-wasi-threads while building threaded runtime
Without this in C/CXXFLAG, wasi-libcxx fails to build against wasi-libc
0.20240412, which has moved pthread.h into the wasm32-wasi-threads
directory (so wasi-libcxx fails with a pthread.h not found error).

I am trying to unblock the Loongarch builder which built wasi-libcxx
after the wasi-libc upgrade (other archs did not do this), so pkgrel
can be bumped later on once it is agreed that this is the proper solution.