community/nawk: upgrade to 20240623
Update url. Old url redirects to this one.

Fix check so that it aligns with upstream's instructions.
Previously, when check was run, the tests were comparing the newly
built awk against the build environment's awk (busybox). The intent
is to test the new build against previous to check for regressions.

To remedy this, include source for previous tagged version to test
against. This only builds if check() is run, but is not installed;
skipped if bootstrapping.