community/callaudiod: split callaudiocli into subpackage
callaudiocli works with any implementation of the callaudiod D-Bus API,
so splitting it into its own package allows it to be installed without
callaudiod. The callaudiod package depends on it, so existing installs
of callaudiod will pull it in and be unaffected.

This change also fixes the license field to match upstream's license file.