community/wlroots: make wlroots-dev depend on the origin package
wlroots 0.18.0 is the first release to allow parallel installs of
multiple wlroots versions [1]. To enable this, upstream removed the
/usr/lib/libwlroots.so symlink that other version of wlroots provided.
Therefore, as this symlink is no longer included in the -dev subpackage,
abuild's dependency tracing algorithm fails to identify that the -dev
subpackage depends on the original package for the shared object.

As such, installing `wlroots-dev` does not automatically pull in
`wlroots` and the latter will need to be specified manually. This commit
resolves that by explicitly specifying a dependency on `wlroots` via
`depends_dev`.

This was discovered while upgrading river to a version which
requires wlroots 0.18.0, see !69716 for more details.

[1]: https://gitlab.freedesktop.org/wlroots/wlroots/-/releases/0.18.0#changes-for-packagers