community/timed: remove unnecessary -recursive option
For some reason when qmake-qt5 is given the -recursive the environmental
variables are read at a different time which prevented the soname fix
from working as expected. Removing the option cause it isn't needed
anyway. The build output is 100% the same.

Fixes <250a39d999f> community/timed: fix so:* provider version