community/phosh: add libphosh subpackages
Phosh is making an increasing amount of its functionality available for
consumption via a shared library.

At this time, the main `phosh-session` binary does *not* use this shared
library, though hopefully it will soon enough [1]. So right now this new
libphosh package has no consumers in aports.

phrog [2] is an upcoming project that I intend to submit to aports
imminently, it is a greetd-based greeter that uses libphosh to provide a
nice mobile-oriented greeter.

[1]: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/1100
[2]: https://github.com/samcday/phrog