testing/wabt: Disable link-time optimization
Wabt compiles two static libraries: libwabt.a and libwasm-rt-impl.a.
When forced to compile with LTO, those libraries end up with LLVM
bytecode files with .o extension, causing linking to fail in a lot of
cases.

CMake lacks sufficient support for LTO to compile static libraries
correctly, so I think the best solution for now is to disable it.