main/gcc: add bootstrap support for gcc-gdc
Since GCC 12, gcc-gdc now requires a working D compiler to bootstrap.

Archs that don't currently have GDC need to have it cross-compiled.

The changes here are mainly in 3 places:

(1) `if $LANG_D` in the global scope is mostly a copy of what
    `if $LANG_ADA` does for cross-bootstrapping. For GDC, we
    only build the dynamic libgphobos when CBUILD == CTARGET.

(2) In build(), the changes are to accommodate bootstrapping with
    a static libphobos, and the fact that parallel make fails when
    cross-compiling the D standard library (this doesn't seem to be
    the case when compiling natively).

(3) Lastly, the changes in the gdc() split function: gcc is suffixed
    with $_target in depends= for the GDC cross-compiler to depend on
    the GCC cross-compiler, libgphobos is only added to subpackages=
    when CBUILD == CTARGET, gcc-gdc-bootstrap and /usr/bin/gdc are
    only provided when CHOST == CTARGET, and finally, when
    CHOST != CTARGET, the libraries are placed in /usr/$CTARGET/lib