community/calls: explicitly depend on sofia-sip-linphone
apk installs dependenies by shared object providers, so while this
package is built against sofia-sip-linphone-dev, when apk traces
dependencies, it looks for the shared objects calls depends on, and
these are provided by both sofia-sip and sofia-sip-linphone. As such, it
may end up picking the former, which results in broken behaviour.

To fix this, explicitly depend on sofia-sip-linphone.