main/alpine-conf: install under /usr
The files installed by alpine-conf are not required to boot a device,
but to configure and set it up. They are mostly useful in boot-media,
where there is a single partition, and no risk of them not being
available. So move them outside the root to the /usr hierarchy

Relates https://gitlab.alpinelinux.org/alpine/tsc/-/issues/73