community/cgmanager: remove
The repository has been archived upstream for 5 years, and main page[1]
has a note that says:

Note

CGManager was used by default with LXC in Ubuntu since April 2014 and
then by other distributions as they started needing working
unprivileged containers.

It has now been deprecated in favor of the CGroup namespace in recent
Linux kernels. On older kernels, LXCFS still offers a cgroupfs
emulation that can be used instead of CGManager and is more widely
compatible with existing userspace.

[1] https://linuxcontainers.org/cgmanager/

So let's drop the package, should certainly not be something we need
in alpine 3.21