community/kwin: apply fix for the Librem 5
The Librem 5 is currently broken with Plasma. Interestingly enough the
device reports 3 displays even though it really only as 2 (internal and
USB-C out) and KWin picks the wrong. Then it tries to do a buffer copy
between... the same device...

This patch makes KWin use the right display. Figured out on Akademy with
a KWin dev, yay for conferences!