community/plasma-desktop-meta: depend on power-profiles-daemon
It's an optional but recommended dependency of powerdevil with which it
allows the user to switch between 3 power profiles (conservative,
balanced, performance) which impacts performance and battery life.
Since powerdevil functions perfectly fine without it I'm putting the
dependency here so people can choose to not install it if they so
desire.