testing/resources: correct build profile
Resources has a default profile and a development profile.
The development profile is meant to be used for nightly
development builds, and the default profile is meant to
be used for releases and distribution.

With no flags specified, Resources counterintuitively
uses the "development" profile as the default, meaning
we get the development styling, icon, and app ID instead
of the regular look and ID.

This commit ensures we use the right profile.