main/glib: suppress output of glib-compile-schemas trigger
The output of compile-schemas, such as warnings about bad dconf paths,
is not intended to be read by end users and is meant only for developers.

Supress only these warnings if /bin/sh is a symlink to busybox.
Otherwise discard all warnings, for example if the user has dash-binsh
or yash-binsh installed.