community/brightnessctl: restrict openrc service to backlights
The application brightnessctl can control two different classes of
sysfs devices, "backlight" and "leds". When it searches for devices
it first goes thtrough the "backlight" class and thereafter through
the "leds" class. If no class or device is specified by the command
line, it acts on the first device it has found.

Using the openrc service script on boot works fine if there is a
backlight available. However, if there is no backlight, brightnessctl
starts to mess around with the first LED it finds, causing potentially
strange things.

Therefore exit the openrc service if there is no backlight.