main/openssh: force static linking against utmps and skalibs
Since openssh-9.9_p1-r0, openssh is linked dynamically against skalibs
and statically against utmps (see the dependency tracing output in the
build log). This is not intended as it should be statically linked
against both. This commit adds a more fancy linker command-line to
ensure that this is the case.

Furthermore, in the future it may be desirable to only link sshd-session
against utmps-dev and not the entire thing.