main/gtk+3.0: add patch to fix touch on slow devices
We've run this on laptops in postmarketOS for a very long while,
without remarkable issues. This greatly helps in performance with slow
devices (e.g: SoCs and phones), and wont get accepted upstream due to
it being in feature freeze, and requiring more extensive changes to
input to not be "just a hack". This seems, however, a sensible patch
downstream for an old version of the project