main/perl-<various>: use direct dep to perl-test-simple
perl-test2-suite was merged into perl-test-simple. Depend directly on
the latter to simplify the build time dependency chain.

This is a follow up to commit 74e36977a593 (main/perl-test2-suite:
remove, merged into main/perl-test-simple) and solves a problem when
bootstrapping aports, which will not pull in indirect dependencies in
the build chain.

>>> perl-extutils-cchecker: Analyzing dependencies...
ERROR: unable to select packages:
  .makedepends-perl-extutils-cchecker-20241014.211514:
    masked in: cache
    satisfies: world[.makedepends-perl-extutils-cchecker=20241014.211514]
  perl-test2-suite (no such package):
    required by: .makedepends-perl-extutils-cchecker-20241014.211514[perl-test2-suite]