main/aports-build: bail out if $arch is empty
If '$arch' is empty, the `rsync` invocation to upload the packages for
the current repo targets the wrong directory, resulting in the
repositories for all other architectures to be removed.

Because it's not known yet what caused the `abuild -A` invocation to
fail, we detect and log it and then bail out before it can do any
damage.

See alpine/infra/infra#10831