testing/py3-drf-orjson-renderer: support numpy 2.0
Adds https://github.com/brianjbuck/drf_orjson_renderer/pull/26:

```
As suggested in #24, DjangoNumpyJSONEncoder should no longer be
necessary when using orjson 3.3 or newer. This fixes an issue when this
package is used with numpy 2.0, which has removed np.float_. However,
since we'd be removing existing functionality, I think this has to be a
breaking change. I did not update the version, but I guess this should
probably mean going to 2.0.0.
```