main/postgresql17: disable server tests on 32bit arches
Most of the server tests fail on 32-bit arches with the message
"test process exited with exit code 2" and I don’t know why. I think
that PostgreSQL 17 is currently broken on them, but I don’t have the
time or motivation to investigate it further at this moment (why would
anyone run PostgreSQL on 32-bit...?).

Tracked in #16582