testing/avara: install files to /usr rather than /var
According to the FHS /var is meant for "variable data", a.k.a. files
that change when the application is running and are created by the
program. That's not what files installed by the package manager are.
Instead such, architecture-independent, data files are supposed to be
installed to /usr/share.

Ideally the binary gets installed to /usr/bin rather than /usr/share
because it _is_ architecture dependent, but it seems this game loads it
resource files relative from where the binary is installed, thus moving
it to /usr/bin doesn't work. So sadly we'll have to install this binary
to /usr/share as well.