community/libusbsio: upgrade to 2.1.13
Also resolves `404 Not Found` error while fetching the source archive.

The source URL has been temporarily changed to use a versioned tarball
cached on dev.alpinelinux.org due to the upstream versioned URL
returning a 404 error. A search results page[1] still shows the 2.1.11
version, but the Download link redirects to a 404 page. The product page
URL also no longer lists downloads with the versioned link, but a
`libusbsio-Host-library.zip` for the 2.1.13 version.

[1]: https://www.nxp.com/search?keyword=libusbsio&start=0