main/busybox: enable cut -F
cut -f is limited to only single character delimiters which means it's not
usable to handle the output of most shell commands.

cut -F is often times a simpler alternative to awk. When the number of
fields is unknown it can be a major simplification compared to awk.

As far as I have seen this doesn't cause a size increase cause the code
has not been optimized out by the compiler.