community/geoclue: use BeaconDB as default Wi-Fi data source
BeaconDB is an Ichnaea-compatible compatible service that's intended to
replace Mozilla's now defunct Location Services (MLS). It's using
crowd-sourced data under an open license and is as of
https://codeberg.org/beacondb/beacondb/issues/5#issuecomment-2504311
available to be used by distributions like ours.

Any source is better than none I'd say and BeaconDB looks to be a great
one.